+++ platform/src/sdk/sdk.ts
@@ -60,9 +60,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "1.4.1";
-    genVersion = "2.143.2";
-    userAgent = "speakeasy-sdk/typescript 1.4.1 2.143.2 3.0.0 @codat/platform";
+    sdkVersion = "1.4.2";
+    genVersion = "2.144.7";
+    userAgent = "speakeasy-sdk/typescript 1.4.2 2.144.7 3.0.0 @codat/platform";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);